# expenses-service/docs/swagger.yaml
basePath: /api
definitions:
  models.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  models.Expense:
    properties:
      amount:
        example: 100.5
        type: number
      category_id:
        example: 1
        type: integer
      company_id:
        example: 1
        type: integer
      created_at:
        example: "2024-03-20T00:00:00Z"
        type: string
      date:
        example: "2024-03-20T00:00:00Z"
        type: string
      description:
        example: Office supplies
        type: string
      id:
        example: 1
        type: integer
      registered_at:
        type: string
      updated_at:
        example: "2024-03-20T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  models.ExpenseCreateRequest:
    properties:
      amount:
        example: 100.5
        type: number
      category_id:
        example: 1
        type: integer
      company_id:
        example: 1
        type: integer
      date:
        example: "2024-03-20T00:00:00Z"
        type: string
      description:
        example: Office supplies
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - amount
    - category_id
    - company_id
    - date
    - description
    - user_id
    type: object
  models.ExpenseListResponse:
    properties:
      expenses:
        items:
          $ref: '#/definitions/models.Expense'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  models.ExpenseUpdateRequest:
    properties:
      amount:
        example: 100.5
        type: number
      category_id:
        example: 1
        type: integer
      date:
        example: "2024-03-20T00:00:00Z"
        type: string
      description:
        example: Office supplies
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A microservice for managing expenses
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Expense Service API
  version: "1.0"
paths:
  /companies/{company_id}/expenses:
    get:
      consumes:
      - application/json
      description: Get a paginated list of expenses for a company
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExpenseListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List expenses
      tags:
      - expenses
    post:
      consumes:
      - application/json
      description: Create a new expense
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: Expense data
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/models.ExpenseCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create expense
      tags:
      - expenses
  /companies/{company_id}/expenses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an expense
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete expense
      tags:
      - expenses
    get:
      consumes:
      - application/json
      description: Get a single expense by ID
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get expense
      tags:
      - expenses
    put:
      consumes:
      - application/json
      description: Update an existing expense
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      - description: Expense data
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/models.ExpenseUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update expense
      tags:
      - expenses
  /companies/{company_id}/expenses/by-date:
    get:
      consumes:
      - application/json
      description: Get a paginated list of expenses for a company within a date range
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Category ID
        in: query
        name: category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExpenseListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List expenses by date range
      tags:
      - expenses
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"